pull tab components into own dir
