+++ src/components/search/result-modal/tabs/index.js
@@ -1,0 +1,4 @@
+export * from './knowledge-graphs'
+export * from './overview'
+export * from './studies'
+export * from './tranql'
