+++ src/components/search/result-modal/tabs/knowledge-graphs.js
@@ -1,0 +1,13 @@
+import { Space, Typography } from 'antd'
+import { KnowledgeGraphs } from '../../'
+
+const { Title } = Typography
+
+export const KnowledgeGraphsTab = ({ graphs }) => {
+  return (
+    <Space direction="vertical">
+      <Title level={ 4 }>Knowledge Graphs</Title>
+      <KnowledgeGraphs graphs={ graphs } />
+    </Space>    
+  )
+}
